finished "feature" - constants/     - renamed from ambiguous 'db'
- all
    - renamed things to fit pip8
