+++ util/generic_response_generator.py
@@ -1,5 +1,5 @@
 from random import choice
-# TODO move lists to db
+# TODO move lists to constants
 # TODO switch to TB_random
 generic_answers = [
     "אני לא מאמין ששאלת את זה",
